fix: the two off-platform CI breaks - LLP64 narrowing and an unused fn
Both are documented traps in CLAUDE.md that I walked into anyway, and
both were invisible to every local gcc/clang build.

WINDOWS (error C2397, sharedstr.h:157). `JitProbe::strobj_data` was
`long` and is initialised from a POINTER DIFFERENCE. Windows is LLP64 -
`long` is 32 bits there - so the braced init from __int64 is a NARROWING
CONVERSION and MSVC rejects it outright. `ptrdiff_t` now. Introduced by
9880182 (the #95 SharedStr probe) and green on gcc/clang for a day
before the Windows lane saw it, which is the whole reason that lane
exists.

MACOS (-Werror,-Wunused-function, tests.cpp). `fwd_opcode_name` - the
helper for the lever-A coverage ratchet - sat OUTSIDE
`#if ML_JIT_SUPPORTED`, so off-platform its only caller compiles out and
it becomes an unused function. CLAUDE.md names this exact failure, for
this exact reason; the fix is the guard it should have had.

VERIFIED with the procedure CLAUDE.md prescribes for this class rather
than by pushing and hoping: jit.h's platform test flipped to `#if 0` and
built with BOTH g++ and clang++ (clean, and -rt still 1921/1921 with the
tier compiled out), plus a CMake build, since CMake is what CI runs and
`make` alone passed every broken configuration here.

A scan for sibling LLP64 traps (`long` fields in headers initialised
from pointer arithmetic) found none.

+++ src/sharedstr.h
@@ -147,7 +147,12 @@ public:
         const void *len;
         const void *slice;
         const void *obj;
-        long strobj_data;
+        /* ⛔ ptrdiff_t, NOT long: this is initialised from a POINTER
+         * DIFFERENCE, and Windows is LLP64 - `long` is 32 bits there,
+         * so a braced init from __int64 is a NARROWING CONVERSION and
+         * MSVC rejects it outright (error C2397). Green on gcc/clang
+         * for a day before the Windows lane saw it. */
+        ptrdiff_t strobj_data;
         /* the object the offset is INTO, so the emitter can bound it
          * without needing StrObj (which is private) visible there */
         size_t strobj_size;
